Moved solver-related header files into a separate klee/Solver/ directory.
