+++ include/klee/Solver/Solver.h
@@ -11,8 +11,8 @@
 #define KLEE_SOLVER_H
 
 #include "klee/Expr/Expr.h"
-#include "klee/SolverCmdLine.h"
 #include "klee/Internal/System/Time.h"
+#include "klee/Solver/SolverCmdLine.h"
 
 #include <vector>
