+++ lib/Core/UserSearcher.cpp
@@ -13,9 +13,8 @@
 #include "Executor.h"
 
 #include "klee/Internal/Support/ErrorHandling.h"
-#include "klee/SolverCmdLine.h"
 #include "klee/MergeHandler.h"
-
+#include "klee/Solver/SolverCmdLine.h"
 
 #include "llvm/Support/CommandLine.h"
