+++ lib/Solver/AssignmentValidatingSolver.cpp
@@ -9,8 +9,8 @@
 
 #include "klee/Expr/Constraints.h"
 #include "klee/Expr/Assignment.h"
-#include "klee/Solver.h"
-#include "klee/SolverImpl.h"
+#include "klee/Solver/Solver.h"
+#include "klee/Solver/SolverImpl.h"
 
 #include <vector>
