+++ lib/Solver/CachingSolver.cpp
@@ -8,14 +8,13 @@
 //===----------------------------------------------------------------------===//
 
 
-#include "klee/Solver.h"
+#include "klee/Solver/Solver.h"
 
 #include "klee/Expr/Constraints.h"
 #include "klee/Expr/Expr.h"
-#include "klee/IncompleteSolver.h"
-#include "klee/SolverImpl.h"
-
-#include "klee/SolverStats.h"
+#include "klee/Solver/IncompleteSolver.h"
+#include "klee/Solver/SolverImpl.h"
+#include "klee/Solver/SolverStats.h"
 
 #include <ciso646>
 #ifdef _LIBCPP_VERSION
