+++ lib/Solver/CoreSolver.cpp
@@ -10,11 +10,14 @@
 #include "STPSolver.h"
 #include "Z3Solver.h"
 #include "MetaSMTSolver.h"
-#include "klee/SolverCmdLine.h"
+
+#include "klee/Solver/SolverCmdLine.h"
 #include "klee/Internal/Support/ErrorHandling.h"
-#include "klee/Solver.h"
+#include "klee/Solver/Solver.h"
+
 #include "llvm/Support/ErrorHandling.h"
 #include "llvm/Support/raw_ostream.h"
+
 #include <string>
 
 namespace klee {
