+++ lib/Solver/MetaSMTSolver.cpp
@@ -16,8 +16,8 @@
 #include "klee/Expr/Constraints.h"
 #include "klee/Expr/ExprUtil.h"
 #include "klee/Internal/Support/ErrorHandling.h"
-#include "klee/Solver.h"
-#include "klee/SolverImpl.h"
+#include "klee/Solver/Solver.h"
+#include "klee/Solver/SolverImpl.h"
 
 #include "llvm/Support/ErrorHandling.h"
