+++ lib/Solver/STPBuilder.cpp
@@ -11,8 +11,8 @@
 #include "STPBuilder.h"
 
 #include "klee/Expr/Expr.h"
-#include "klee/Solver.h"
-#include "klee/SolverStats.h"
+#include "klee/Solver/Solver.h"
+#include "klee/Solver/SolverStats.h"
 #include "klee/util/Bits.h"
 
 #include "ConstantDivision.h"
