+++ lib/Solver/Solver.cpp
@@ -7,10 +7,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "klee/Solver.h"
+#include "klee/Solver/Solver.h"
 
 #include "klee/Expr/Constraints.h"
-#include "klee/SolverImpl.h"
+#include "klee/Solver/SolverImpl.h"
 
 using namespace klee;
