+++ lib/Solver/SolverImpl.cpp
@@ -7,8 +7,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "klee/Solver.h"
-#include "klee/SolverImpl.h"
+#include "klee/Solver/Solver.h"
+#include "klee/Solver/SolverImpl.h"
 
 using namespace klee;
