+++ lib/Solver/ValidatingSolver.cpp
@@ -8,8 +8,8 @@
 //===----------------------------------------------------------------------===//
 
 #include "klee/Expr/Constraints.h"
-#include "klee/Solver.h"
-#include "klee/SolverImpl.h"
+#include "klee/Solver/Solver.h"
+#include "klee/Solver/SolverImpl.h"
 
 #include <vector>
