+++ lib/Solver/Z3Builder.cpp
@@ -12,8 +12,8 @@
 
 #include "klee/Expr/Expr.h"
 #include "klee/Internal/Support/ErrorHandling.h"
-#include "klee/Solver.h"
-#include "klee/SolverStats.h"
+#include "klee/Solver/Solver.h"
+#include "klee/Solver/SolverStats.h"
 #include "klee/util/Bits.h"
 
 #include "llvm/ADT/StringExtras.h"
